remove trace-specific schematic symbol selector
