+++ lib/components/primitive-components/SchematicSymbol/SchematicSymbol.ts
@@ -1,18 +1,11 @@
 import { schematicSymbolProps } from "@tscircuit/props"
 import { PrimitiveComponent } from "../../base-components/PrimitiveComponent"
-import type { Port } from "../Port"
-import {
-  type MappedSchematicSymbolPort,
-  SchematicSymbol_doInitialPortMatching,
-  getConnectionNamesForSymbolPort,
-} from "./SchematicSymbol_doInitialPortMatching"
 import { SchematicSymbol_doInitialSchematicComponentRender } from "./SchematicSymbol_doInitialSchematicComponentRender"
 
 export class SchematicSymbol extends PrimitiveComponent<
   typeof schematicSymbolProps
 > {
   isSchematicPrimitive = true
-  mappedPorts: MappedSchematicSymbolPort[] = []
 
   get config() {
     return {
@@ -30,18 +23,6 @@ export class SchematicSymbol extends PrimitiveComponent<
       are_pins_interchangeable: false,
     })
     this.source_component_id = sourceComponent.source_component_id
-  }
-
-  doInitialPortMatching(): void {
-    SchematicSymbol_doInitialPortMatching(this)
-  }
-
-  _resolveMappedPort(portName: string): Port | null {
-    return (
-      this.mappedPorts.find(({ symbolPort }) =>
-        getConnectionNamesForSymbolPort(symbolPort.labels).includes(portName),
-      )?.referencedPort ?? null
-    )
   }
 
   doInitialSchematicComponentRender(): void {
