+++ lib/components/primitive-components/SchematicSymbol/SchematicSymbol_doInitialSchematicComponentRender.ts
@@ -2,10 +2,31 @@ import type { SchematicPort } from "circuit-json"
 import { getRelativeDirection } from "lib/utils/get-relative-direction"
 import { getPinNumberFromLabels } from "lib/utils/getPortFromHints"
 import { symbols } from "schematic-symbols"
+import type { Port } from "../Port"
 import type { SchematicSymbol } from "./SchematicSymbol"
 
 const getDisplayPinLabel = (labels: string[]): string | undefined =>
   labels.find((label) => !/^(pin)?\d+$/.test(label))
+
+const getConnectionNamesForSymbolPort = (labels: string[]): string[] => {
+  const connectionNames = new Set(labels)
+  const pinNumber = getPinNumberFromLabels(labels)
+  if (pinNumber) connectionNames.add(`pin${pinNumber}`)
+  return [...connectionNames]
+}
+
+const getSingleConnectionTarget = (
+  connectionName: string,
+  target: string | readonly string[],
+): string => {
+  const targets = Array.isArray(target) ? target : [target]
+  if (targets.length !== 1) {
+    throw new Error(
+      `SchematicSymbol connection "${connectionName}" must select exactly one physical port`,
+    )
+  }
+  return targets[0]
+}
 
 export const SchematicSymbol_doInitialSchematicComponentRender = (
   schematicSymbol: SchematicSymbol,
@@ -20,12 +41,100 @@ export const SchematicSymbol_doInitialSchematicComponentRender = (
     throw new Error(`No schematic symbol found for "${symbolName}"`)
   }
 
-  const referencedSourcePortIds = new Map(
-    schematicSymbol.mappedPorts.map(({ symbolPort, referencedPort }) => [
-      symbolPort,
-      referencedPort.source_port_id!,
-    ]),
-  )
+  const { chipRef, connections } = schematicSymbol._parsedProps
+  if (chipRef && !connections) {
+    db.source_property_ignored_warning.insert({
+      source_component_id: schematicSymbol.source_component_id!,
+      property_name: "chipRef",
+      message: `${schematicSymbol.getString()} has chipRef "${chipRef}" without connections. chipRef will be ignored.`,
+      error_type: "source_property_ignored_warning",
+      subcircuit_id: schematicSymbol.getSubcircuit().subcircuit_id ?? undefined,
+    })
+  }
+  if (connections && !chipRef) {
+    db.source_property_ignored_warning.insert({
+      source_component_id: schematicSymbol.source_component_id!,
+      property_name: "connections",
+      message: `${schematicSymbol.getString()} has connections without chipRef. connections will be ignored.`,
+      error_type: "source_property_ignored_warning",
+      subcircuit_id: schematicSymbol.getSubcircuit().subcircuit_id ?? undefined,
+    })
+  }
+
+  const effectiveChipRef = connections ? chipRef : undefined
+  const effectiveConnections = chipRef ? connections : undefined
+  const referencedChip = effectiveChipRef
+    ? schematicSymbol.getSubcircuit().selectOne(effectiveChipRef)
+    : null
+  if (effectiveChipRef && !referencedChip?.source_component_id) {
+    throw new Error(
+      `Could not resolve chipRef "${effectiveChipRef}" for ${schematicSymbol.getString()}`,
+    )
+  }
+
+  const usedConnectionNames = new Set<string>()
+  const referencedSourcePortIds = new Map<
+    (typeof symbol.ports)[number],
+    string
+  >()
+
+  if (effectiveConnections && referencedChip?.source_component_id) {
+    for (const symbolPort of symbol.ports) {
+      const connectionNames = getConnectionNamesForSymbolPort(symbolPort.labels)
+      const matchingConnections = Object.entries(effectiveConnections).filter(
+        ([connectionName]) => connectionNames.includes(connectionName),
+      )
+      if (matchingConnections.length === 0) {
+        throw new Error(
+          `${schematicSymbol.getString()} is missing an explicit connection for symbol port "${symbolPort.labels.join("/")}"`,
+        )
+      }
+      if (matchingConnections.length > 1) {
+        throw new Error(
+          `${schematicSymbol.getString()} has multiple connections for symbol port "${symbolPort.labels.join("/")}"`,
+        )
+      }
+
+      const [connectionName, connectionTarget] = matchingConnections[0]
+      usedConnectionNames.add(connectionName)
+      const targetSelector = getSingleConnectionTarget(
+        connectionName,
+        connectionTarget,
+      )
+      const referencedPort = schematicSymbol
+        .getSubcircuit()
+        .selectOne(targetSelector, { type: "port" }) as Port | null
+
+      if (!referencedPort?.source_port_id) {
+        throw new Error(
+          `Could not resolve connection "${connectionName}" target "${targetSelector}" for ${schematicSymbol.getString()}`,
+        )
+      }
+
+      const referencedSourcePort = db.source_port.get(
+        referencedPort.source_port_id,
+      )
+      if (
+        referencedSourcePort?.source_component_id !==
+        referencedChip.source_component_id
+      ) {
+        throw new Error(
+          `Connection "${connectionName}" target "${targetSelector}" does not belong to chipRef "${effectiveChipRef}"`,
+        )
+      }
+
+      referencedSourcePortIds.set(symbolPort, referencedPort.source_port_id)
+    }
+
+    const unusedConnectionNames = Object.keys(effectiveConnections).filter(
+      (connectionName) => !usedConnectionNames.has(connectionName),
+    )
+    if (unusedConnectionNames.length > 0) {
+      throw new Error(
+        `${schematicSymbol.getString()} has connections that do not match symbol ports: ${unusedConnectionNames.join(", ")}`,
+      )
+    }
+  }
 
   const center = schematicSymbol._getGlobalSchematicPositionBeforeLayout()
   const schematicSheetId = schematicSymbol._resolveSchematicSheetId()
