+++ tests/components/primitive-components/schematic-symbol-connections-mosfet.test.tsx
@@ -1,5 +1,4 @@
 import { expect, test } from "bun:test"
-import type { Port } from "lib/components"
 import { getTestFixture } from "tests/fixtures/get-test-fixture"
 
 test("schematicsymbol maps MOSFET symbol ports to a chip for traces", async () => {
@@ -42,7 +41,7 @@ test("schematicsymbol maps MOSFET symbol ports to a chip for traces", async () =
         schX={2}
         schY={0.55}
       />
-      <trace from=".R1 > .pin2" to="Q1A.gate" />
+      <trace from=".R1 > .pin2" to=".Q1 > .G1" />
       <trace from=".R2 > .pin1" to=".Q1 > .D1" />
 
       <chip
@@ -66,22 +65,6 @@ test("schematicsymbol maps MOSFET symbol ports to a chip for traces", async () =
   )
 
   await circuit.renderUntilSettled()
-
-  const physicalGatePort = circuit.selectOne<Port>("Q1.G1", { type: "port" })
-  const resistorOutputPort = circuit.selectOne<Port>("R1.pin2", {
-    type: "port",
-  })
-  const inputSourceTrace = circuit.db.source_trace
-    .list()
-    .find((trace) =>
-      trace.connected_source_port_ids.includes(
-        resistorOutputPort?.source_port_id ?? "",
-      ),
-    )
-
-  expect(inputSourceTrace?.connected_source_port_ids).toContain(
-    physicalGatePort?.source_port_id,
-  )
 
   await expect(circuit).toMatchStackedSchematicSnapshot(import.meta.path)
   await expect(circuit).toMatchPcbSnapshot(import.meta.path)
